+++ src/table/column.rs
@@ -43,7 +43,7 @@ pub trait IntoColumnDef {
 /// | Bit                   | bit               | bit                         | N/A                          |
 /// | VarBit                | bit               | varbit                      | N/A                          |
 /// | Boolean               | bool              | bool                        | boolean                      |
-/// | Money                 | money             | money                       | real_money                   |
+/// | Money                 | decimal           | money                       | real_money                   |
 /// | Json                  | json              | json                        | json_text                    |
 /// | JsonBinary            | json              | jsonb                       | jsonb_text                   |
 /// | Uuid                  | binary(16)        | uuid                        | uuid_text                    |
@@ -75,7 +75,7 @@ pub enum ColumnType {
     TimestampWithTimeZone,
     Time,
     Date,
-    Year(Option<MySqlYear>),
+    Year,
     Interval(Option<PgInterval>, Option<u32>),
     Binary(u32),
     VarBinary(StringLen),
@@ -114,7 +114,6 @@ impl PartialEq for ColumnType {
             (Self::Char(l0), Self::Char(r0)) => l0 == r0,
             (Self::String(l0), Self::String(r0)) => l0 == r0,
             (Self::Decimal(l0), Self::Decimal(r0)) => l0 == r0,
-            (Self::Year(l0), Self::Year(r0)) => l0 == r0,
             (Self::Interval(l0, l1), Self::Interval(r0, r1)) => l0 == r0 && l1 == r1,
             (Self::Binary(l0), Self::Binary(r0)) => l0 == r0,
             (Self::VarBinary(l0), Self::VarBinary(r0)) => l0 == r0,
@@ -151,6 +150,17 @@ impl ColumnType {
     {
         ColumnType::Custom(Alias::new(ty).into_iden())
     }
+
+    pub fn string(length: Option<u32>) -> ColumnType {
+        match length {
+            Some(s) => ColumnType::String(StringLen::N(s)),
+            None => ColumnType::String(StringLen::None),
+        }
+    }
+
+    pub fn var_binary(length: u32) -> ColumnType {
+        ColumnType::VarBinary(StringLen::N(length))
+    }
 }
 
 /// All column specification keywords
@@ -184,13 +194,6 @@ pub enum PgInterval {
     HourToMinute,
     HourToSecond,
     MinuteToSecond,
-}
-
-// All MySQL year type field length sizes
-#[derive(Debug, Clone, Eq, PartialEq)]
-pub enum MySqlYear {
-    Two,
-    Four,
 }
 
 impl ColumnDef {
@@ -475,8 +478,8 @@ impl ColumnDef {
 
     /// Set column type as year
     /// Only MySQL supports year
-    pub fn year(&mut self, length: Option<MySqlYear>) -> &mut Self {
-        self.types = Some(ColumnType::Year(length));
+    pub fn year(&mut self) -> &mut Self {
+        self.types = Some(ColumnType::Year);
         self
     }
 
@@ -507,6 +510,12 @@ impl ColumnDef {
     pub fn varbit(&mut self, length: u32) -> &mut Self {
         self.types = Some(ColumnType::VarBit(length));
         self
+    }
+
+    #[cfg(feature = "backend-mysql")]
+    /// Set column type as blob. MySQL only.
+    pub fn blob(&mut self) -> &mut Self {
+        self.custom(crate::extension::mysql::MySqlType::Blob)
     }
 
     /// Set column type as boolean